Improve return check on function GetGameDir_s
Now returns a size_t representing 0 on success and the minimum size of the buffer to contain the full file path on failure.

+++ op2ext.h
@@ -22,8 +22,8 @@ extern "C" {
 
 // Retrieves the current absolute directory of the Outpost 2 executable with a trailing slash. 
 // If bufferSize is smaller than required to copy entire path, buffer is provided as much of path as possible.
-// Returns 0 on success.
-OP2EXT_API int GetGameDir_s(char* buffer, size_t bufferSize);
+// Returns 0 on success. Returns the required minimum size of the buffer on failure.
+OP2EXT_API size_t GetGameDir_s(char* buffer, size_t bufferSize);
 
 
 // DEPRECATED as of version 2.0.0. Use GetGameDir_s instead.
